requirements.txt
carla
numpy
opencv-python
pygame
tqdm
Pillow
PyYAML